refactor(DID): split DID verification steps functions
