+++ src/inspectors/did/index.ts
@@ -1,0 +1,3 @@
+import domain from '../../domain';
+
+export const baseError = domain.i18n.getText('errors', 'identityErrorBaseMessage');
